brd-testcase-automation/requirements.txt
altair==5.5.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
asn1crypto==1.4.0
attrs==25.3.0
Babel==2.9.1
bcrypt==4.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
cached-property==2.0.1
cachetools==5.5.2
cbor2==5.4.2
certifi==2025.8.3
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.37.0
colorama==0.4.6
contourpy==1.3.2
coverage==7.10.7
cryptography==46.0.3
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
Django==5.1.4
dnspython==2.4.2
docopt==0.6.2
docutils==0.17
email-validator==2.1.0.post1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
flask-mailman==1.0.0
flask-marshmallow==1.3.0
Flask-Principal==0.4.0
flask-restx==1.3.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.59.2
freezegun==1.1.0
geoip2==2.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
greenlet==3.0.1
gspread==5.12.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
libsass==0.20.1
lxml==4.8.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.2
matplotlib==3.10.6
maxminddb==2.8.2
narwhals==2.3.0
num2words==0.5.10
numpy==1.26.4
oauthlib==3.3.1
ofxparse==0.21
openai==2.6.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.4
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
polib==1.1.1
protobuf==6.32.0
psutil==5.9.0
psycopg2==2.9.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.0
pyOpenSSL==21.0.0
pyparsing==3.2.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pyserial==3.5
PySocks==1.7.1
pytest==7.4.3
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ldap==3.4.0
python-stdnum==1.17
pytz==2025.2
pyusb==1.2.1
qrcode==7.3.1
referencing==0.36.2
reportlab==3.6.8
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rjsmin==1.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
selenium==4.35.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.23
sqlparse==0.5.3
streamlit==1.49.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vobject==0.9.6.1
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0
zipp==3.17.0
